atualizando tipo do dado

+++ labs/lab03/SimpleSample/src/main/java/example/SimpleSample.java
@@ -101,7 +101,7 @@ public class SimpleSample extends ChaincodeBase {
 		}		
 		if(valA>valFrom)
 			return "{\"Error\":\"Insufficient asset holding value for requested transfer amount \"}";
-		float juliotax=1;
+		int juliotax=1;
 		valFrom = valFrom-valA-juliotax;
 		valTo = valTo+valA;
 		System.out.println("Transfer "+fromName+">"+toName+" am='"+am+"' new values='"+valFrom+"','"+ valTo+"'");
